Write a test for Env::Global
The class Env provides the global namespace in Node. The global
namespace has many properties, such as setImmediate. require() is
one of a handful objects that are available for use in Node modules,
which are not found within the global namespace. This test checks
that the object returned from Env::Global both has setImmediate as
a property and does not have require() as a property.